Formatting in Coords & ClientChunkSize, fixed AutoReply
